webrtc/RTCPeerConnection-peerIdentity.html

<!doctype html>
<meta charset=utf-8>
<title>RTCPeerConnection.prototype.getIdentityAsssertion</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="identity-helper.js"></script>
<script>
  'use strict';

  // Test is based on the following editor draft:
  // https://w3c.github.io/webrtc-pc/archives/20170605/webrtc.html

  /*
    9.6.  RTCPeerConnection Interface Extensions
      partial interface RTCPeerConnection {
        void               setIdentityProvider(DOMString provider,
                                               optional RTCIdentityProviderOptions options);
        Promise<DOMString> getIdentityAssertion();
        readonly attribute Promise<RTCIdentityAssertion> peerIdentity;
      };

      dictionary RTCIdentityProviderOptions {
        DOMString protocol = "default";
        DOMString usernameHint;
        DOMString peerIdentity;
      };

      [Constructor(DOMString idp, DOMString name)]
      interface RTCIdentityAssertion {
        attribute DOMString idp;
        attribute DOMString name;
      };

    4.3.2. setRemoteDescription
      If an a=identity attribute is present in the session description, the browser
      validates the identity assertion..

      If the "peerIdentity" configuration is applied to the RTCPeerConnection, this
      establishes a target peer identity of the provided value. Alternatively, if the
      RTCPeerConnection has previously authenticated the identity of the peer (that
      is, there is a current value for peerIdentity ), then this also establishes a
      target peer identity.

      The target peer identity cannot be changed once set. Once set, if a different
      value is provided, the user agent MUST reject the returned promise with a newly
      created InvalidModificationError and abort this operation. The RTCPeerConnection
      MUST be closed if the validated peer identity does not match the target peer
      identity.

      If there is no target peer identity, then setRemoteDescription does not await the
      completion of identity validation.

    9.4.  Verifying Identity Assertions
      The identity assertion request process involves the following asynchronous steps:
        1.  The RTCPeerConnection awaits any prior identity validation. Only one identity
            validation can run at a time for an RTCPeerConnection. This can happen because
            the resolution of setRemoteDescription is not blocked by identity validation
            unless there is a target peer identity.

        2.  The RTCPeerConnection loads the identity assertion from the session description
            and decodes the base64 value, then parses the resulting JSON. The idp parameter
            of the resulting dictionary contains a domain and an optional protocol value
            that identifies the IdP, as described in [RTCWEB-SECURITY-ARCH].

        3.  The RTCPeerConnection instantiates the identified IdP as described in 9.1.1
            Identity Provider Selection and 9.2 Registering an IdP Proxy. If the IdP
            cannot be loaded, instantiated or the IdP proxy is not registered, this
            process fails.

        4.  The RTCPeerConnection invokes the validateAssertion method registered by the IdP.

            The assertion parameter is taken from the decoded identity assertion. The origin
            parameter contains the origin of the script that calls the RTCPeerConnection
            method that triggers this behavior.

        5.  The IdP proxy returns a promise and performs the validation process asynchronously.

            The IdP proxy verifies the identity assertion using whatever means necessary.
            Depending on the authentication protocol this could involve interacting with the
            IdP server.

        6.  If the IdP proxy produces an error or returns a promise that does not resolve
            to a valid RTCIdentityValidationResult (see 9.5 IdP Error Handling), then
            identity validation fails.

        7.  Once the assertion is successfully verified, the IdP proxy resolves the promise
            with an RTCIdentityValidationResult containing the validated identity and the
            original contents that are the payload of the assertion.

        8.  The RTCPeerConnection decodes the contents and validates that it contains a
            fingerprint value for every a=fingerprint attribute in the session description.
            This ensures that the certificate used by the remote peer for communications
            is covered by the identity assertion.

        10. The RTCPeerConnection resolves the peerIdentity attribute with a new instance
            of RTCIdentityAssertion that includes the IdP domain and peer identity.

        11. The user agent MAY display identity information to a user in its UI. Any user
            identity information that is displayed in this fashion MUST use a mechanism that
            cannot be spoofed by content.

      If identity validation fails, the peerIdentity promise is rejected with a newly
      created OperationError.

      If identity validation fails and there is a target peer identity for the
      RTCPeerConnection, the promise returned by setRemoteDescription MUST be rejected
      with the same DOMException.

      If identity validation fails and there is no a target peer identity, the value of
      the peerIdentity MUST be set to a new, unresolved promise instance. This permits
      the use of renegotiation (or a subsequent answer, if the session description was
      a provisional answer) to resolve or reject the identity.

    9.5.  IdP Error Handling
      - A RTCPeerConnection might be configured with an identity provider, but loading of
        the IdP URI fails. Any procedure that attempts to invoke such an identity provider
        and cannot load the URI fails with an RTCError with errorDetail set to
        "idp-load-failure" and the httpRequestStatusCode attribute of the error set to the
        HTTP status code of the response.

      - If the IdP loads fails due to the TLS certificate used for the HTTPS connection not
        being trusted, it fails with an RTCError with errorDetail set to "idp-tls-failure".
        This typically happens when the IdP uses certificate pinning and an intermediary
        such as an enterprise firewall has intercepted the TLS connection.

      - If the script loaded from the identity provider is not valid JavaScript or does not
        implement the correct interfaces, it causes an IdP failure with an RTCError with
        errorDetail set to "idp-bad-script-failure".

      - An apparently valid identity provider might fail in several ways.

        If the IdP token has expired, then the IdP MUST fail with an RTCError with
        errorDetail set to "idp-token-expired".

        If the IdP token is not valid, then the IdP MUST fail with an RTCError with
        errorDetail set to "idp-token-invalid".

        If an identity provider throws an exception or returns a promise that is
        ultimately rejected, then the procedure that depends on the IdP MUST also fail.
        These types of errors will cause an IdP failure with an RTCError with errorDetail
        set to "idp-execution-failure".

      - The user agent SHOULD limit the time that it allows for an IdP to 15 seconds.
        This includes both the loading of the IdP proxy and the identity assertion
        generation or validation. Failure to do so potentially causes the corresponding
        operation to take an indefinite amount of time. This timer can be cancelled when
        the IdP proxy produces a response. Expiration of this timer cases an IdP failure
        with an RTCError with errorDetail set to "idp-timeout".

      - If the identity provider requires the user to login, the operation will fail
        RTCError with errorDetail set to "idp-need-login" and the idpLoginUrl attribute
        of the error set to the URL that can be used to login.

      - Even when the IdP proxy produces a positive result, the procedure that uses this
        information might still fail. Additional validation of a RTCIdentityValidationResult
        value is still necessary. The procedure for validation of identity assertions
        describes additional steps that are required to successfully validate the output
        of the IdP proxy.
   */

  const host = window.location.host;
  const domain = window.location.hostname;

  const alice = `alice@${domain}`;
  const bob = `bob@${domain}`;

  promise_test(() => {
    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection();

    pc1.setIdentityProvider(host, {
      protocol: 'idp-test.js',
      usernameHint: alice
    });

    return pc1.createOffer()
    .then(offer => pc2.setRemoteDescription(offer))
    .then(() => pc2.peerIdentity)
    .then(identityAssertion => {
      const { idp, name } = identityAssertion;
      assert_equals(idp, host, `Expect IdP domain to be ${domain}`);
      assert_equals(identityAssertion, alice, `Expect validated identity to be ${alice}`);
    });
  }, 'setRemoteDescription() on offer with a=identity should establish peerIdentity');

  /*
    4.3.2. setRemoteDescription
      The target peer identity cannot be changed once set. Once set, if a different
      value is provided, the user agent MUST reject the returned promise with a newly
      created InvalidModificationError and abort this operation. The RTCPeerConnection
      MUST be closed if the validated peer identity does not match the target peer
      identity.
   */
  promise_test(t => {
    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection({ peerIdentity: bob });

    pc1.setIdentityProvider(host, {
      protocol: 'idp-test.js',
      usernameHint: alice
    });

    return pc1.createOffer()
    .then(offer =>
      promise_rejects(t, 'InvalidModificationError',
        pc2.setRemoteDescription(offer)))
      .then(() => {
        assert_true(pc2.signalingState, 'closed',
          'Expect peer connection to be closed after mismatch peer identity');
      });
  }, 'setRemoteDescription() on offer with a=identity that resolve to value different from target peer identity should reject with InvalidModificationError');

  promise_test(t => {
    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection({ peerIdentity: alice });

    pc1.setIdentityProvider(host, {
      protocol: 'idp-test.js?returnContents=bogus',
      usernameHint: alice
    });

    const peerIdentityPromise = pc2.peerIdentity;

    return pc1.createOffer()
    .then(offer => Promise.all([
      promise_rejects(t, new RTCError(),
        pc2.setRemoteDescription(offer)),
      promise_rejects(t, 'OperationError',
        peerIdentityPromise)
    ]);

  }, 'setRemoteDescription() with peerIdentity set and with IdP proxy that return validationAssertion with mismatch contents should reject with RTCError');

  /*
    9.4.  Verifying Identity Assertions
      9.  The RTCPeerConnection validates that the domain portion of the identity matches
          the domain of the IdP as described in [RTCWEB-SECURITY-ARCH]. If this check
          fails then the identity validation fails.
   */
  promise_test(t => {
    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection();

    pc1.setIdentityProvider(host, {
      protocol: 'idp-test.js?action=validate-throw-error',
      usernameHint: alice
    });

    const peerIdentityPromise = pc2.peerIdentity;

    return pc1.createOffer()
    .then(offer => pc2.setRemoteDescription(offer))
    .then(() =>
      promise_rejects(t, 'OperationError',
        peerIdentityPromise))
    .then(() => {
      const peerIdentityPromise2 = pc2.peerIdentity

      pc1.setIdentityProvider(host, {
        protocol: 'idp-test.js',
        usernameHint: alice
      });

      return pc1.createOffer()
      .then(offer => pc2.setRemoteDescription(offer))
      .then(peerIdentityPromise2)
      .then(identityAssertion => {
        const { idp, name } = identityAssertion;
        assert_equals(idp, host, `Expect IdP domain to be ${domain}`);
        assert_equals(identityAssertion, alice, `Expect validated identity to be ${alice}`);
      });
    });

  }, 'IdP failure with no target peer identity should have setRemoteDescription succeed and replace pc.peerIdentity with a new promise');

</script>
